requirements.txt
pandas
pinecone-client
sentence-transformers
python-dotenv
